[fix] #267 가게 상세 프로필 조회 리팩토링
